Finalizing - Image Upload

+++ app/Http/Controllers/PostsController.php
@@ -3,11 +3,22 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -46,14 +57,32 @@ class PostsController extends Controller
         //validation
         $this->validate($request, [
           'title' => 'required',
-          'body' => 'required'
+          'body' => 'required',
+          'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        //Handle File Upload
+        if($request->hasFile('cover_image')) {
+          //get filename with extension
+          $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+          //get just filename
+          $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+          //get just extension
+          $fileExt = $request->file('cover_image')->getClientOriginalExtension();
+          //filename to Store
+          $fileNameToStore = $filename.'_'.time().'.'.$fileExt;
+          //Upload image
+          $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+          $fileNameToStore = 'no_image.jpg';
+        }
 
         //create posts
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -82,6 +111,11 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+
+        //check correct user_id
+        if(auth()->user()->id !== $post->user_id){
+          return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -100,10 +134,27 @@ class PostsController extends Controller
           'body' => 'required'
         ]);
 
+        //Handle File Upload
+        if($request->hasFile('cover_image')) {
+          //get filename with extension
+          $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+          //get just filename
+          $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+          //get just extension
+          $fileExt = $request->file('cover_image')->getClientOriginalExtension();
+          //filename to Store
+          $fileNameToStore = $filename.'_'.time().'.'.$fileExt;
+          //Upload image
+          $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
         //Update post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')) {
+          $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -119,6 +170,17 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+
+        //check correct user_id
+        if(auth()->user()->id !== $post->user_id){
+          return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        if($post->cover_image != 'no_image.jpg'){
+          //delete image
+          Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
